WebService/requirements.txt
Django==1.10.4
django-appconf==1.0.2
django-filter==1.0.1
django-imagekit==3.3
djangorestframework==3.5.3
djangorestframework-jwt==1.9.0
Markdown==2.6.7
pilkit==1.1.13
Pillow==3.4.2
pkg-resources==0.0.0
PyJWT==1.4.2
six==1.10.0
